Move old tests to process_message_spec and correct them to make it's possible to nuke the old implementation
